Cleanup: Remove uneccessary Javadoc comments, rename class.
